ダウンロードリンク用画面をDownloadFromLinkModalに分離
