+++ frontend/src/features/home/components/MainContents.tsx
@@ -1,22 +1,19 @@
-import { LoadingSpinner } from '@/components/LoadingSpinner';
 import { useSelectBox } from '@/hooks/useSelectBox';
-import { notifyState } from '@/stores';
 import { relativePathSlicer } from '@/utils/slice';
 import { Box, Button, List, Stack, Typography } from '@mui/material';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { MdArrowBack } from 'react-icons/md';
-import { useSetRecoilState } from 'recoil';
 import { useSelector } from '../../../hooks/useSelector';
 import { useDownload } from '../api/download';
 import { useSendRequest } from '../api/sendRequest';
 import { Uploads } from '../api/upload';
 import { Storage } from '../types/storage';
 import DirpathNavigation from './main/DirpathNavigation';
+import DownloadFromLinkModal from './main/DownloadFromLinkModal';
 import EmptyDirDisplay from './main/EmptyDirDisplay';
 import FilePathList from './main/FilePathList';
 import ProgressBars from './main/ProgressBars';
-import { SelectList } from './SelectList';
 
 type MainContentsProps = {
   filepaths: Storage['filepaths'];
@@ -26,20 +23,6 @@ type MainContentsProps = {
   uploads: Uploads;
   important?: boolean;
   moveDir: (path: string) => Promise<void>;
-};
-
-const modalStyle = {
-  position: 'absolute' as 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  bgcolor: 'background.paper',
-  p: 5,
-  px: 10,
 };
 
 const selectSortValues = ['昇順', '降順'];
@@ -64,8 +47,7 @@ export const MainContents = ({
   // セレクトボックス用hooks
   const [SortSelectForm, selectedSortValue] = useSelectBox('sort', selectSortValues);
   const [PrioritySelectForm, setctedPriorityValue] = useSelectBox('priority', selectPriorityValues);
-  // recoil
-  const setNotify = useSetRecoilState(notifyState);
+  // ダウンロードリンク関連
   const [downloadFromLink, setDownloadFromLink] = useState<boolean>(false);
   const [downloadSelectedArray, setDownloadSelectedArray] = useState<
     { name: string; type: 'dir' | 'file' }[]
@@ -76,6 +58,7 @@ export const MainContents = ({
       | 'dir'
       | 'file',
   }));
+  // ディレクトリパス関連
   const prevPath = currentdir.slice(0, currentdir.lastIndexOf('/'));
   const relativePath = relativePathSlicer(currentdir, baseDir);
   const dirs = relativePath.split('/');
@@ -101,55 +84,14 @@ export const MainContents = ({
   // ダウンロードリンクを開いた際の画面
   if (downloadFromLink) {
     return (
-      <Box
-        sx={{
-          position: 'absolute',
-          width: '100%',
-          height: '100vh',
-          bgcolor: 'rgba(0,0,0,0.8)',
-          zIndex: 100,
-        }}
-      >
-        <ProgressBars
-          downloadProgresses={downloadProgresses}
-          downloadCancelMutation={downloadCancelMutation}
-        />
-        <Box sx={modalStyle}>
-          <Stack sx={{ py: 2, px: 10 }} spacing={2} alignItems="center">
-            <div>以下をダウンロードしますか？</div>
-            <SelectList selects={downloadSelectedArray} />
-            <Stack direction="row" spacing={2}>
-              <Button
-                sx={{ whiteSpace: 'nowrap' }}
-                size="medium"
-                variant="contained"
-                onClick={async () => {
-                  setNotify({ severity: 'info', text: 'ダウンロード後, 移動します' });
-                  await downloadMutation
-                    .mutateAsync({ path: currentdir, targets: downloadSelectedArray })
-                    .finally(async () => {
-                      setDownloadFromLink(false);
-                      await router.push(`/?path=${router.query.path}`);
-                    });
-                }}
-              >
-                {downloadMutation.isLoading && (
-                  <LoadingSpinner size="sm" sx={{ color: '#fff', mr: 1 }} />
-                )}
-                ダウンロード
-              </Button>
-              <Button
-                onClick={async () => {
-                  setDownloadFromLink(false);
-                  await router.push(`/?path=${router.query.path}`);
-                }}
-              >
-                閉じる
-              </Button>
-            </Stack>
-          </Stack>
-        </Box>
-      </Box>
+      <DownloadFromLinkModal
+        currentDir={currentdir}
+        downloadSelectedArray={downloadSelectedArray}
+        downloadProgresses={downloadProgresses}
+        downloadMutation={downloadMutation}
+        downloadCancelMutation={downloadCancelMutation}
+        setDownloadFromLink={setDownloadFromLink}
+      />
     );
   }
